Move VoiceChannel.Effect to VoiceChannelEffect
